% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaN.R
\name{apply_delta_doc}
\alias{apply_delta_doc}
\title{Function analyze delta N per document}
\usage{
apply_delta_doc(filename, D, V, FUN = sum, ...)
}
\arguments{
\item{filename}{Filename of delta N file}

\item{D}{The number of documents to read in.}

\item{V}{The vocabulary size.}

\item{FUN}{Function to use to calculate for each Delta N.}

\item{...}{Further arguments to send to FUN.}
}
\value{
A list with a list of results from function.
}
\description{
Apply an arbitrary function to a delta N matrix per document.
}
\examples{
\dontrun{
  filename <- "Runs/Run2014-11-12--10_31_02/delta_n/DeltaNs_noDocs_1499_vocab_12255_iter_1000.BINARY" 
  test <- apply_delta_doc(filename, D = 100, V = 12255, FUN=function(X) sum(abs(X))/2)
}
}

