% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ineff.R
\name{calculate_inefficiency}
\alias{calculate_inefficiency}
\title{Calculate inefficiency of theta and phi}
\usage{
calculate_inefficiency(bin_dir, settings)
}
\arguments{
\item{bin_dir}{Search path to directory with binary files.}

\item{settings}{list with settings to use (verbose, top_no_words, test_doc, alpha_prior, beta_prior)}
}
\value{
A list with params, ineff_phi and ineff_theta
}
\description{
Reads data and calculates inefficiencies for Theta and Phi
}

