% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_rareword.R
\name{calculate_rareword}
\alias{calculate_rareword}
\title{Calculate the rare word limit based on proportion in corpus}
\usage{
calculate_rareword(corpus, prop = 0.001, have_enough_memory = TRUE)
}
\arguments{
\item{corpus}{Corpus on the form used in Mallet}

\item{proportion}{Proportion of tokens that will be removed for a given rare word limit}
}
\description{
Function to calculate the rare word limit for a given corpus
}
\examples{
corpus <- "datasets/enron_random.txt"
calculate_rareword(corpus)
}

