% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phi_theta.R
\name{check_phi}
\alias{check_phi}
\title{Function to do checks on the Phi matrix}
\usage{
check_phi(phi_mat)
}
\arguments{
\item{phi_mat}{A Phi matrix}
}
\description{
Function to do checks on the Phi matrix
}

