% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phi_theta.R
\name{check_theta}
\alias{check_theta}
\title{Function to do checks on the Theta matrix}
\usage{
check_theta(theta_mat)
}
\arguments{
\item{theta_mat}{A theta matrix}
}
\description{
Function to do checks on the Theta matrix
}

