% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsc_adlda_mode.R
\name{create_nsc_runs_adlda_mode}
\alias{create_nsc_runs_adlda_mode}
\title{Function to create bash files, config files and NSC Runs for AD-LDA mode test}
\usage{
create_nsc_runs_adlda_mode(path_to_project = "", java_seeds, mem, K,
  data_paths, cores, mail_to, samplers)
}
\description{
Function to generate bash scripts and config files to run at NSC.
 The jar file assumes to be localed at the part_to_project.
 The purpose of the experiment is to see if there is differences in modes
 at different cores.
 
 @inheritParams create_nsc_runs_ineff
 @param cores Vector with the number of cores to use
 
 @examples
 create_nsc_runs_adlda_mode(java_seeds = 20150326, 
                         cores = c(8, 4, 2, 1), 
                         K = c(20, 100), 
                         data_paths = c("datasets/enron.txt", "datasets/nips.txt"),
                         mail_to = "leif.jonsson[at]ericsson.com")
                         
 @export
}

