% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nsc_runs.R
\name{create_nsc_runs_ineff}
\alias{create_nsc_runs_ineff}
\title{Function to create bash files, config files and NSC Runs}
\usage{
create_nsc_runs_ineff(path_to_project = "", java_seeds, K, data_paths,
  R_in_parallel = FALSE, cores = 8, mail_to, delete_files = FALSE)
}
\description{
Function to generate bash scripts and config files to run at NSC.
 The jar file assumes to be localed at the part_to_project.
 
 @param path_to_project Path to the project 
 @param java_seeds The different runs to use (one run for each seed)
 @param K The number of topic (vector)
 @param data_paths The datasets to use (vector) as path
 @param R_in_parallel Should the postprocessing in R use paralell processing?
 @param cores The number of cores to use (default is 8)
 @param mail_to e-mail adress to use for jobs
 @param delete_files Should the files be deleted after run?
 
 @export
}

