% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsc_prior.R
\name{create_nsc_runs_prior}
\alias{create_nsc_runs_prior}
\title{Function to create bash files, config files and NSC Runs for the PRIOR experiment}
\usage{
create_nsc_runs_prior(path_to_project = "", java_seeds, K, data_paths, alpha,
  beta, cores_no = 8, h = 5, mem, samplers, mail_to)
}
\description{
Function to generate bash scripts and config files to run at NSC.
 The jar file assumes to be localed at the part_to_project.
 
 @inheritParams create_nsc_runs_ineff
 @param cores Vector with the number of cores to use
 @param samplers Vector with the different samplers to use
 @param huge_node Should the huge node be used?
 
 @examples
 create_nsc_runs_prior(java_seeds = 20150326, 
                         K = c(100, 200), 
                         data_paths = "datasets/enron.txt",
                         samplers = c("spalias", "adlda"),
                         alpha = c(0.1, 0.01),
                         beta = c(0.1, 0.01),
                         mail_to = "leif.jonsson[at]ericsson.com")
                         
 @export
}

