% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsc_pubmed.R
\name{create_nsc_runs_pubmed}
\alias{create_nsc_runs_pubmed}
\title{Function to create bash files, config files and NSC Runs for the PUBMED experiment}
\usage{
create_nsc_runs_pubmed(path_to_project = "", folder = "pubmed", java_seeds,
  K, data_paths, max_doc_buf_size = NULL, rare_word_limits = NULL,
  tfidf_vocab_size = NULL, cores = NULL, samplers, mail_to,
  huge_node = TRUE, mem, time_limit_h = NULL, alpha_beta_prior = NULL,
  account, call = "PCPLDA.jar cc.mallet.topics.tui.ParallelLDA",
  iterations = 1000)
}
\description{
Function to generate bash scripts and config files to run at NSC.
 The jar file assumes to be localed at the part_to_project.
 
 @inheritParams create_nsc_runs_ineff
 @param cores Vector with the number of cores to use
 @param samplers Vector with the different samplers to use
 @param huge_node Should the huge node be used?
}
\examples{
create_nsc_runs_pubmed(java_seeds = 20150326, 
                         cores = c(64, 32), 
                         K = c(200, 400, 800), 
                         data_paths = "datasets/pubmed.txt",
                         samplers = c("spalias", "adlda"),
                         mail_to = "leif.jonsson[at]ericsson.com")

}

