% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsc_rare_word.R
\name{create_nsc_runs_rarewords}
\alias{create_nsc_runs_rarewords}
\title{Function to create bash files, config files and NSC Runs for Spalias experiment}
\usage{
create_nsc_runs_rarewords(path_to_project = "", java_seeds, K, rare_words,
  data_paths, mail_to, cores = 8)
}
\description{
Function to generate bash scripts and config files to run at NSC.
 The jar file assumes to be localed at the part_to_project.
 
 @inheritParams create_nsc_runs_ineff
 @param cores Number of cores to use

 @examples
 create_nsc_runs_rarewords(java_seeds=20150326 + 0:2, 
                           K=c(20,100), 
                           data_paths = c("datasets/enron_random.txt", "datasets/nips_random.txt"), 
                           rare_words = list(c(14,43,104,193), c(12,30,62,108)), 
                           mail_to="mons.magnusson[at]gmail.com")
             
 @export
}

