% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nsc_runs.R
\name{create_nsc_runs_rs}
\alias{create_nsc_runs_rs}
\title{Function to create bash files, config files and NSC Runs for RS experiment}
\usage{
create_nsc_runs_rs(path_to_project = "", java_seeds, K, data_paths, mail_to,
  cores = 8, sub_samplers = c("standard", "delta_n", "token_mass"),
  full_phi_it = 1, percentage_phi = 1, percentage_z = 1)
}
\description{
Function to generate bash scripts and config files to run at NSC.
 The jar file assumes to be localed at the part_to_project.
 
 @inheritParams create_nsc_runs_ineff
 @param cores Vector with the number of cores to use
 @param samplers Vector with the different samplers to use
 @param samplers Vector with the different samplers to use
 
 @examples
 create_nsc_runs_rs(java_seeds = 20150326 + (0:5), 
                    K = c(50, 100), 
                    data_paths = c("datasets/enron.txt", "datasets/nips.txt"),
                    sub_samplers = c("full", "token_mass", "delta_n"),
                    full_phi_it = c(5,10),
                    percentage_phi=c(0.01, 0.05, 0.2, 0.5),
                    percentage_z=c(1, 0.5, 0.2),
                    mail_to = "mons.magnusson[at]gmail.com, leif.jonsson[at]ericsson.com")
                         
 @export
}

