% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_nsc_runs.R
\name{create_nsc_runs_spalias}
\alias{create_nsc_runs_spalias}
\title{Function to create bash files, config files and NSC Runs for Spalias experiment}
\usage{
create_nsc_runs_spalias(path_to_project = "", java_seeds, K, data_paths,
  mail_to, cores = 8)
}
\description{
Function to generate bash scripts and config files to run at NSC.
 The jar file assumes to be localed at the part_to_project.
 
 @inheritParams create_nsc_runs_ineff
 @param cores Number of cores to use

 @examples
   create_nsc_runs_spalias(java_seeds=2000 + 1:5,
                           K=c(20,100),
                           data_paths = c("datasets/enron.txt", "datasets/nips.txt"),
                           mail_to="mons.magnusson[at]gmail.com"
                           )          
 @export
}

