% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsc_variable_selection.R
\name{create_nsc_runs_vs}
\alias{create_nsc_runs_vs}
\title{Function to create bash files, config files and NSC Runs for VS-prop experiment}
\usage{
create_nsc_runs_vs(path_to_project = "", folder, java_seeds, K, data_paths, h,
  mail_to, mem, cores = 16, pis, account, iterations = 5000)
}
\arguments{
\item{cores}{Vector with the number of cores to use}

\item{samplers}{Vector with the different samplers to use}

\item{samplers}{Vector with the different samplers to use}
}
\description{
Function to generate runs for random scan tokens proportional to size.
}

