% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_wiki_corpus.R
\name{file_to_doc_vector}
\alias{file_to_doc_vector}
\title{Parse file content}
\usage{
file_to_doc_vector(raw_file)
}
\arguments{
\item{raw_file}{A raw file read in from wikipedia corpus}
}
\description{
Parse a read raw file in tagged.en files to one doc per element vector.
}

