% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaN.R
\name{get_bin_size_ints}
\alias{get_bin_size_ints}
\title{Get the binary file element size for a given binary file with integers}
\usage{
get_bin_size_ints(filename)
}
\arguments{
\item{filename}{File to get the number of elements from}
}
\description{
Get the binary file element size for a given binary file with integers
}

