% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_from_dir.R
\name{get_config_from_dir_recursive}
\alias{get_config_from_dir_recursive}
\title{Function to return all configs from a folder with runsuites}
\usage{
get_config_from_dir_recursive(path)
}
\arguments{
\item{path}{Path to folder with runsuites}
}
\description{
Return the java setup files from any given run using tgs.runnables.InefficencyExperiment
}
\examples{
\dontrun{
  path <- "/Users/manma97/Desktop/Runs/"
  get_config_from_dir_recursive(path)
}
}

