% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_from_dir.R
\name{get_experiment_details_from_path}
\alias{get_experiment_details_from_path}
\title{Get parameters of interest for the actual experiment}
\usage{
get_experiment_details_from_path(bin_dir)
}
\arguments{
\item{bin_dir}{Path to a directory with binary files calculated with tgs.runnables.InefficencyExperiment}
}
\value{
A list containing the parameters:
  experiment_no, algo_type, iterations and run
}
\description{
Compute the parameters of interest from a given path
}

