% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_from_dir.R
\name{get_likelihoods_from_dir}
\alias{get_likelihoods_from_dir}
\title{Function to get likelihoods from files}
\usage{
get_likelihoods_from_dir(path)
}
\arguments{
\item{path}{Path to folder with runsuites}
}
\description{
Return list with loglikelihood calculations
}
\examples{
\dontrun{
  path <- "/Users/manma97/Desktop/Runs/"
  get_config_from_dir_recursive(path)
}
}

