% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_likelihoods_from_paths.R
\name{get_likelihoods_from_paths}
\alias{get_likelihoods_from_paths}
\title{Function to get likelihoods from file paths}
\usage{
get_likelihoods_from_paths(paths, head = c("iter", "loglik", "timestamp"))
}
\arguments{
\item{paths}{Vector with paths to likelihood files}

\item{head}{Variable names in file - if null header is in file}
}
\description{
Return list with loglikelihood calculations
}
\examples{
\dontrun{
  path <- "/Users/manma97/Desktop/Runs/"
  get_config_from_dir_recursive(path)
}
}

