% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_from_dir.R
\name{get_param_from_config}
\alias{get_param_from_config}
\title{Function to parse cfg_configs}
\usage{
get_param_from_config(cfg_file, pattern)
}
\arguments{
\item{cfg_file}{The cfg_file to get the parameter from}

\item{pattern}{The parameter that should be return as regexp}
}
\description{
Function to parse cfg_configs
}

