% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_from_dir.R
\name{get_param_from_convergence}
\alias{get_param_from_convergence}
\title{Function to parse cfg_configs}
\usage{
get_param_from_convergence(conv_files, pattern)
}
\arguments{
\item{conv_files}{The convergence files to get the parameter from}

\item{pattern}{The parameter that should be return as regexp}
}
\description{
Function to parse cfg_configs
}

