% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_from_dir.R
\name{get_param_from_dir}
\alias{get_param_from_dir}
\title{Function that get all parameters of interest from a binary directory}
\usage{
get_param_from_dir(bin_dir)
}
\arguments{
\item{bin_dir}{Path to a directory with binary files calculated with tgs.runnables.InefficencyExperiment}
}
\value{
A list containing the parameters:
  experiment_no, V, K, D, algo_type, iterations, run_suite
}
\description{
Compute the parameters of interest from binary files (in a binary directory)
}

