% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results_from_paths.R
\name{get_results_from_paths}
\alias{get_results_from_paths}
\title{Function to get stats from stats.txt file and likelihood.txt paths}
\usage{
get_results_from_paths(lik_paths, stat_paths, all_cores = FALSE,
  runtime_unit = NULL)
}
\arguments{
\item{lik_paths}{Likelihood paths}

\item{stat_paths}{Stat paths}

\item{all_cores}{Should all results be included (count update times and z update times)}

\item{timediff_unit}{If choosen (secs or mins) the timediff is calculated. Default it is not calculated (ie NULL).}
}
\description{
Return list with results (times etc)
}

