% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_param_from_dir.R
\name{get_runsuite_from_config}
\alias{get_runsuite_from_config}
\title{Calculate the Runsuite based on the config}
\usage{
get_runsuite_from_config(project_path, config)
}
\arguments{
\item{project_path}{path to folder with all RunSuites}

\item{config}{The config to get the runsuite for (ex. "nips_10")}
}
\description{
Calculate the Runsuite based on the config
}
\details{
It checks the subconfig name from the convergence txt-file (Active subconfig)
}

