% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineff_factor.R
\name{get_top_D_documents_param_mat}
\alias{get_top_D_documents_param_mat}
\title{Function to calculate theta for a subset of documents}
\usage{
get_top_D_documents_param_mat(filenames, D, K, alpha_prior, verbose = TRUE)
}
\arguments{
\item{filenames}{M binary files to use (search paths)}

\item{D}{the number of documents to use (top D documents)}

\item{K}{the number of topics}

\item{alpha_prior}{the alpha prior for M}

\item{verbose}{Should the calculations be verbose?}
}
\value{
A list with matrices of same size, stored with iterations as rows.
}
\description{
Calculates theta from an M matrix.
}

