% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineff_factor.R
\name{get_top_phi_from_N_param_mat}
\alias{get_top_phi_from_N_param_mat}
\title{Function to calculate phi from a N matrix}
\usage{
get_top_phi_from_N_param_mat(filenames, K, V, beta_prior, top_no_words = NULL,
  verbose = TRUE)
}
\arguments{
\item{filenames}{filenames (vector) to use for calculations}

\item{K}{The number of topics}

\item{V}{The vocabulary size}

\item{beta_prior}{the prior for phi}

\item{top_no_words}{the number of top words in topics to use.}

\item{verbose}{Should the run be verbose?}
}
\value{
A list with matrices of same size, stored with iterations as rows.
}
\description{
Calculates phi for top words from an N matrix.
}

