% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineff_factor.R
\name{get_top_phi_param_mat}
\alias{get_top_phi_param_mat}
\title{Function to calculate top words}
\usage{
get_top_phi_param_mat(filenames, K, V, top_no_words = NULL, verbose = TRUE)
}
\arguments{
\item{filenames}{filenames (vector) to use for calculations}

\item{K}{The number of topics}

\item{V}{The vocabulary size}

\item{top_no_words}{the number of top words in topics to use.}

\item{verbose}{Should the run be verbose?}
}
\value{
A list of length K with matrices of size T * top_no_words, stored with iterations as rows.
}
\description{
Calculates the top words from a param_matrix based on mean value of phi_i
}
\examples{
\dontrun{
  topic_top_word_list <- 
    get_top_topics_param_mat(filenames, K=20, V = 12127, top_no_words = 5, verbose = TRUE)
}
}

