% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inefficiency_factors.R
\name{inefficiency_factor}
\alias{inefficiency_factor}
\title{Inefficiency factor (v1)}
\usage{
inefficiency_factor(X)
}
\arguments{
\item{X}{chain to calculate inefficiency for.}
}
\description{
Inefficiency factor (v1)
}

