% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inefficiency_factors.R
\name{inefficiency_factor_v2}
\alias{inefficiency_factor_v2}
\title{Inefficiency factor (v2)}
\usage{
inefficiency_factor_v2(X, lag.max)
}
\arguments{
\item{X}{chain to calculate inefficiency for.}
}
\description{
Inefficiency factor (v2)
}
\details{
Implementation of tau_a
}
\references{
http://stats.stackexchange.com/questions/66369/definition-of-autocorrelation-time-for-effective-sample-size
}

