% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineff_factor.R, R/inefficiency_factors.R
\name{inefficiency_factors}
\alias{inefficiency_factors}
\title{Calculate inefficiency factors (IF)}
\usage{
inefficiency_factors(mat_list, in_parallel = FALSE)

inefficiency_factors(mat_list, in_parallel = FALSE)
}
\arguments{
\item{mat_list}{A list of matrices}

\item{in_parallel}{Should the calculation use multiple cores. Default is FALSE.}

\item{x}{object of class \code{PCLDA_param_list}}

\item{r}{length of sum (see details).}
}
\value{
A list of vectors with IFs.

A matrix of the same size with inefficiency factors per parameter
}
\description{
Calculates IF for all paramaters of an \code{PCLDA_param_list}
  object.

Calculate inefficiency factors from list of matrices
}
\details{
The autocovariance can be calculated in the following way:
  \deqn{\tau_{a}=1+2\cdot\sum_{k=1}^{\infty}\rho_{k}}
}
\examples{
\dontrun{
  inefficiency_factors(x = x, r = 10)
}
}
\references{
\url{http://stats.stackexchange.com/questions/66369/definition-of-autocorrelation-time-for-effective-sample-size}
}

