% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsc_utils.R
\name{num_time_to_nsc_format}
\alias{num_time_to_nsc_format}
\title{Convert numeric number of hours to NSC format}
\usage{
num_time_to_nsc_format(h_num)
}
\arguments{
\item{h_num}{}
}
\description{
Convert numeric number of hours to NSC format
}
\details{
Round up to closest 15 minutes
}
\examples{
num_time_to_nsc_format(0.1)
num_time_to_nsc_format(9.99)
num_time_to_nsc_format(23.45)
}

