% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_data_by_topic_proportions.R
\name{order_data_by_topic_proportions}
\alias{order_data_by_topic_proportions}
\title{Sort corpus by topic proportions}
\usage{
order_data_by_topic_proportions(input_crp, output_sort_crp, output_strat_crp,
  topic_prop, batches)
}
\arguments{
\item{input_crp}{Path to corpus in Mallet format}

\item{output_sort_crp}{Path to output ordered file (by topics)}

\item{output_strat_crp}{Path to output ordered file (by topics) and then equally distributed}

\item{topic_prop}{Path to topic proportions output by PartiallyCollapse doc_topic_mean_filename config argument.}
}
\description{
Sort corpus in order of topics proportions (in order by topics)
}

