% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_wiki_corpus.R
\name{parse_wiki_corpus}
\alias{parse_wiki_corpus}
\title{Parse wiki corpus to Mallet format}
\usage{
parse_wiki_corpus(wiki_folder, tmp_folder = tempdir(), mallet_file)
}
\arguments{
\item{wiki_folder}{The folder from}

\item{mallet_file}{Filename to printout to Mallet format}
}
\description{
This function will parse the wikipedia corpus into Mallet format. The file that can be parsed is
the tagged wikipedia data (Wikipedia corpus v.1.0) found at: 
http://www.cs.upc.edu/~nlp/wikicorpus/
}
\examples{
\dontrun{
  wiki_folder <- "datasets/tagged.en/"
  tmp_folder <- "datasets/tmp/"
  mallet_file <- "datasets/wiki.txt"
  parse_wiki_corpus(wiki_folder = wiki_folder, tmp_folder=tmp_folder, mallet_file = mallet_file)
  }

}

