% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_likelihoods.R
\name{plot_likelihoods}
\alias{plot_likelihoods}
\title{Plot likelihood graphs}
\usage{
plot_likelihoods(file_paths, group_names = NULL, subset = NULL)
}
\arguments{
\item{file_paths}{File paths to the likelihood txt files to plot in the same graph}

\item{group_names}{The name for each file path (needs to be the same length as \code{file_paths}).}

\item{subset}{Numeric vector containing the subset of iterations to use.}
}
\value{
ggplot2 object.
}
\description{
Plots the likelihood files generated by the Java code
}

