% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineff_factor.R
\name{r_dirichlet}
\alias{r_dirichlet}
\title{r_dirichlet}
\usage{
r_dirichlet(n, alpha)
}
\arguments{
\item{n}{number of draws.}

\item{alpha}{matrix or vector}
}
\description{
Random draws from a dirichlet distribution
}

