% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_order_corpus.R
\name{random_order_corpus}
\alias{random_order_corpus}
\title{Function to randomly reorder documents in corpus file}
\usage{
random_order_corpus(input_file, output_file, seed = 4711,
  have_enough_memory = TRUE)
}
\arguments{
\item{input_file}{File to reorder}

\item{output_file}{Reordered file}

\item{seed}{seed to use for random reorder}
}
\description{
Function that reads a corpus file, reorder the documents randomly and printout a new randomly ordered corpus file.
}
\examples{
\dontrun{
  input_file <- "datasets/enron.txt" # 39860
  output_file <- "datasets/enron_test1.txt" 
  random_order_corpus(input_file = "datasets/enron.txt", output_file = "datasets/enron_test1.txt", have_enough_memory = FALSE)
  }
}

