% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaN.R
\name{read_delta_corpus_matrix}
\alias{read_delta_corpus_matrix}
\title{Function to read in delta file to one full delta Matrix (for the whole corpus)}
\usage{
read_delta_corpus_matrix(filename, D, V)
}
\arguments{
\item{filename}{Filename of delta N file}

\item{D}{The number of documents to read in.}

\item{V}{The vocabulary size.}
}
\value{
A sparse matrices of size K * V.
}
\description{
The functions read in deltaN binary files into R and store them as list of spase matrices.
}
\examples{
\dontrun{
  filename <- "Runs/Run2014-11-12--10_31_02/delta_n/DeltaNs_noDocs_1499_vocab_12255_iter_1000.BINARY" 
  test <- read_delta_corpus_matrix(filename, D = 1499, V = 12255)
}
}

