% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltaN.R
\name{read_delta_N}
\alias{read_delta_N}
\title{Function to read in deltaN values}
\usage{
read_delta_N(filename, first, last)
}
\arguments{
\item{filename}{Filename of a delta N file}

\item{first}{First element to read from the binary file.}

\item{last}{Last element to read from the binary file.}
}
\value{
A matrix containing the parameters from \code{first} to \code{last} and 
  with each files as a separate row.
}
\description{
Read in values from \code{first} element to \code{last} element in all
  binary files specified by \code{filenames}.
}

