% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_matrices.R
\name{read_matrices}
\alias{read_matrices}
\title{Function to read in parameter values from binary matricies}
\usage{
read_matrices(filenames, rows, cols, type = c("integer", "numeric"))
}
\arguments{
\item{filenames}{Vector with filenames to binary files to read.}

\item{rows}{The number of rows in the matrices.}

\item{cols}{The number of columns in the matrices.}

\item{type}{Whate type is the binary file ('integer' and 'numeric' is implemented)}
}
\value{
A list of matrices.
}
\description{
Read in values from \code{first} element to \code{last} element in all
  binary files specified by \code{filenames}.
}

