% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ineff_factor.R
\name{read_part_param}
\alias{read_part_param}
\title{Function to read in parameter values}
\usage{
read_part_param(filenames, first, last, type = c("integer", "numeric"))
}
\arguments{
\item{filenames}{Vector with filenames to binary files to read.}

\item{first}{First element to read from the binary file.}

\item{last}{Last element to read from the binary file.}

\item{type}{Whate type is the binary file ('integer' and 'numeric' is implemented)}
}
\value{
A matrix containing the parameters from \code{first} to \code{last} and 
  with each files as a separate row.
}
\description{
Read in values from \code{first} element to \code{last} element in all
  binary files specified by \code{filenames}.
}

