% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_corpus.R
\name{subset_corpus}
\alias{subset_corpus}
\title{Function to randomly reorder documents in corpus file}
\usage{
subset_corpus(input_file, output_file, prop = 0.1, n = NULL)
}
\arguments{
\item{input_file}{File to reorder}

\item{output_file}{Reordered file}

\item{prop}{Proportion of the full corpus}

\item{n}{Number of observations}
}
\description{
Function that reads subsets a corpus (chooses the first part, so use a ranom ordered corpus)
}
\examples{
  input_file <- "datasets/enron.txt" # 39860
  output_file <- "datasets/enron_10.txt" 
  subset_corpus(input_file = input_file, output_file = output_file)
}

